+++ packages/frontend/src/components/common/Filters/FilterInputs/FilterMonthAndYearPicker.tsx
@@ -22,10 +22,9 @@ const FilterMonthAndYearPicker: FC<Props> = ({ value, onChange, ...props }) => {
             onClick={toggle}
             {...props}
             popoverProps={{
-                withArrow: true,
-                withinPortal: false,
                 shadow: 'md',
-                // FIXME: remove this once we migrate off of Blueprint
+                // Month and year picker does not manage its own state properly.
+                // additional props are needed to make it work
                 ...props.popoverProps,
                 opened: isPopoverOpen,
                 onOpen: () => {