+++ packages/frontend/src/components/common/ShareShortLinkButton/ShareShortLinkButton.styles.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import { BigButton } from '../BigButton';
-
-export const ShareLink = styled(BigButton)`
-    width: 40px;
-    height: 40px;
-    margin-left: 12px !important;
-`;